+++ lib/ferrum/network.rb
@@ -11,9 +11,10 @@ module Ferrum
   class Network
     CLEAR_TYPE = %i[traffic cache].freeze
     AUTHORIZE_TYPE = %i[server proxy].freeze
-    RESOURCE_TYPES = %w[Document Stylesheet Image Media Font Script TextTrack
-                        XHR Fetch EventSource WebSocket Manifest
-                        SignedExchange Ping CSPViolationReport Other].freeze
+    REQUEST_STAGES = %i[Request Response].freeze
+    RESOURCE_TYPES = %i[Document Stylesheet Image Media Font Script TextTrack
+                        XHR Fetch Prefetch EventSource WebSocket Manifest
+                        SignedExchange Ping CSPViolationReport Preflight Other].freeze
     AUTHORIZE_BLOCK_MISSING = "Block is missing, call `authorize(...) { |r| r.continue } " \
                               "or subscribe to `on(:request)` events before calling it"
     AUTHORIZE_TYPE_WRONG = ":type should be in #{AUTHORIZE_TYPE}"
@@ -187,11 +188,20 @@ module Ferrum
     #   end
     #   browser.go_to("https://google.com")
     #
-    def intercept(pattern: "*", resource_type: nil)
+    def intercept(pattern: "*", resource_type: nil, request_stage: nil, handle_auth_requests: true)
       pattern = { urlPattern: pattern }
-      pattern[:resourceType] = resource_type if resource_type && RESOURCE_TYPES.include?(resource_type.to_s)
 
-      @page.command("Fetch.enable", handleAuthRequests: true, patterns: [pattern])
+      if resource_type && RESOURCE_TYPES.none?(resource_type.to_sym)
+        raise ArgumentError, "Unknown resource type '#{resource_type}' must be #{RESOURCE_TYPES.join(' | ')}"
+      end
+
+      if request_stage && REQUEST_STAGES.none?(request_stage.to_sym)
+        raise ArgumentError, "Unknown request stage '#{request_stage}' must be #{REQUEST_STAGES.join(' | ')}"
+      end
+
+      pattern[:resourceType] = resource_type if resource_type
+      pattern[:requestStage] = request_stage if request_stage
+      @page.command("Fetch.enable", patterns: [pattern], handleAuthRequests: handle_auth_requests)
     end
 
     #
@@ -323,11 +333,21 @@ module Ferrum
     #
     # @example
     #   browser.network.offline_mode
-    #   browser.go_to("https://github.com/") # => Ferrum::StatusError (Request to https://github.com/ failed to reach
-    #   server, check DNS and server status)
+    #   browser.go_to("https://github.com/")
+    #     # => Request to https://github.com/ failed (net::ERR_INTERNET_DISCONNECTED) (Ferrum::StatusError)
     #
     def offline_mode
       emulate_network_conditions(offline: true, latency: 0, download_throughput: 0, upload_throughput: 0)
+    end
+
+    #
+    # Toggles ignoring cache for each request. If true, cache will not be used.
+    #
+    # @example
+    #   browser.network.cache(disable: true)
+    #
+    def cache(disable:)
+      @page.command("Network.setCacheDisabled", cacheDisabled: disable)
     end
 
     private
@@ -352,6 +372,7 @@ module Ferrum
         if params["redirectResponse"]
           previous_exchange = select(request.id)[-2]
           response = Network::Response.new(@page, params)
+          response.loaded = true
           previous_exchange.response = response
         end
 
@@ -374,8 +395,12 @@ module Ferrum
 
     def subscribe_loading_finished
       @page.on("Network.loadingFinished") do |params|
-        exchange = select(params["requestId"]).last
-        exchange.response.body_size = params["encodedDataLength"] if exchange&.response
+        response = select(params["requestId"]).last&.response
+
+        if response
+          response.loaded = true
+          response.body_size = params["encodedDataLength"]
+        end
       end
     end